+++ libraries/DMM/CHANGELOG.md
@@ -6,13 +6,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2023-02-03
+- add OHM measurements functions
+  - **setReferenceR(ohm), readOhm(times), readKiloOhm(times)**
+- update GitHub actions
+- update license 2023
+- refactor
+- update readme.md
+
+
 ## [0.1.1] - 2022-12-19
 - initial release...
 - renamed initial class to DMM
-- add setGain() and getGain()
-- add getNoise
+- add **setGain()** and **getGain()**
+- add **getNoise()**
 - add examples
-
 
 ## [0.1.0] - eons ago
 - initial version